training/requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
PyYAML
